remove get_resharding_jobs, and its test
get_resharding_jobs() is a compaction-strategy-specific way of deciding
which SSTables to reshard together. The only user that specializes this
is LCS.

The benefit of this is dubious: we try to keep SSTables of the same
level together, but once SSTables from different shards, with different
shapes, get resharded this is essentially lost. The main benefit is
doing many SSTables together, but the new sstable_directory resharder
does that too.

With this patch we are able to remove this special case, and also
the special resharding_compaction_descriptor.

Signed-off-by: Glauber Costa <glauber@scylladb.com>