compaction_manager: carve out a drain method
We want stop() to be callable just once. Having the compaction manager
stopped twice is a potential indication that something is wrong.

Still there are places where we want to stop all ongoing compactions
and prevent new from running - like the drain operation. Today the
only operation that allows for cancellation of all existing compations
is stop(). To unweave this, we will split those two things.

A drain operation is carved out, and it should be safe to be called many
times. The compaction manager is usable after this, and new compactions
can even be sent if it happen to be enabled again (we currently don't)

A stop operation, which includes a drain, will only be allowed once. After
a stop() the compaction_manager object is no longer usable.

Signed-off-by: Glauber Costa <glauber@scylladb.com>