sstables: store data size in foreign_sstable_open_info
In the new version of resharding we'll want to spread SSTables around
the many shards based on their total size. This means we also need to
know the size of each SSTable individually.

We could wrap the foreign_sstable_info around another structure that
keeps track of that, but because this structure exists mostly for
resharding purposes anyway we will just add the data_size to it.

Signed-off-by: Glauber Costa <glauber@scylladb.com>