sstable_directory: add helper to reshape existing unshared sstables
Before moving SSTables to the main directory, we may need to reshape them
into in-strategy. This patch provides helper code that reshapes the SSTables
that are known to be unshared local in the sstable directory, and updates the
sstable directory with the result.

Rehaping can be made more or less aggressive by passing a reshape_threshold,
which is the amount of SSTables reshape can tolerate to consider a particular
slice of the SSTable set offstrategy.

Signed-off-by: Glauber Costa <glauber@scylladb.com>